requirements.txt
aiohttp==3.7.4
ansi2html==1.5.2
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
bagit==1.7.0
bashlex==0.14
CacheControl==0.11.7
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coloredlogs==10.0
cryptography>=3.3.2
cwlref-runner==1.0
cwltool==1.0.20191022103248
decorator==4.4.1
defusedxml==0.7.1
Django==2.2.20
djangorestframework>=3.11.2
Flask==1.0.2
future==0.18.2
humanfriendly==4.18
idna==2.8
importlib-metadata==0.20
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
latexcodec==1.0.6
lockfile==0.12.2
lxml==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
multidict==4.5.2
mypy-extensions==0.4.3
networkx==2.4
oauthlib==3.1.0
packaging==19.1
pathlib2==2.3.5
pluggy==0.12.0
prov==1.5.1
py==1.10.0
pybtex==0.22.2
pycparser==2.19
PyJWT==2.0.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
PySocks==1.6.8
pytest==5.1.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2018.9
PyYAML==5.4
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.2.0
scandir==1.10.0
schema-salad==4.5.20191023134839
selenium==3.141.0
shellescape==3.4.1
simplejson==3.16.0
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.7
Werkzeug==1.0.0
yarl==1.3.0
zipp==0.6.0
